fixed links front page
Hello,

I fixed a very small error presents in the front page of habitica.
ERROR: After an update of the language (in the footer) all the links
presents in the section "player-use-habitica-to-manage", when clicked,
pointed to the top of the page.
NEW BEHAVIOUR: After this modification in front.jade, whenever the user
click the links, it will remain in the same section.